+++ src/components/Main/TotalList/SingleNews/SingleNews.tsx
